skills/firebase-html-deploy/files/requirements.txt
google-cloud-storage>=3.0.0
google-auth>=2.0.0
